+++ Dumper/PriceHistoryEntry.cs
@@ -7,20 +7,20 @@ namespace Dumper
         public PriceHistoryEntry(IDictionary<object, object> entry)
         {
             HistoryDate = (long)entry["historyDate"];
-            LowPrice = (long)entry["lowPrice"];
-            HighPrice = (long)entry["highPrice"];
-            AveragePrice = (long)entry["avgPrice"];
+            LowPrice = (decimal)(double)entry["lowPrice"];
+            HighPrice = (decimal)(double)entry["highPrice"];
+            AveragePrice = (decimal)(double)entry["avgPrice"];
             Volume = (long)entry["volume"];
             Orders = (int)entry["orders"];
         }
 
         internal long HistoryDate { get; private set; }
 
-        internal long LowPrice { get; private set; }
+        internal decimal LowPrice { get; private set; }
 
-        internal long HighPrice { get; private set; }
+        internal decimal HighPrice { get; private set; }
 
-        internal long AveragePrice { get; private set; }
+        internal decimal AveragePrice { get; private set; }
 
         internal long Volume { get; private set; }
